Correct dependencies by moving document-api test to vespaclient-container-plugin
