+++ vespaclient-container-plugin/pom.xml
@@ -77,14 +77,12 @@
       <version>1.8.4</version>
       <scope>test</scope>
     </dependency>
-    <!--
     <dependency>
       <groupId>com.yahoo.vespa</groupId>
       <artifactId>application</artifactId>
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
-    -->
   </dependencies>
   <build>
     <plugins>
