+++ vespaclient-container-plugin/src/test/rest-api-application/services.xml
